Fix TypeScript error while using SvelteKit.
