--- src/lib/ui/svelte/hooks.server.js
@@ -1,9 +1,0 @@
-// To enable o1js for the web, we must set the COOP and COEP headers.
-// See here for more information: https://docs.minaprotocol.com/zkapps/how-to-write-a-zkapp-ui#enabling-coop-and-coep-headers
-/** @type {import('@sveltejs/kit').Handle} */
-export async function handle({ event, resolve }) {
-  const response = await resolve(event);
-  response.headers.set('Cross-Origin-Opener-Policy', 'same-origin');
-  response.headers.set('Cross-Origin-Embedder-Policy', 'require-corp');
-  return response;
-}
