+++ src/lib/ui/svelte/serverHooksSvelte.js
@@ -1,0 +1,24 @@
+export const serverHooksSvelteJavaScript = `
+// To enable o1js for the web, we must set the COOP and COEP headers.
+// See here for more information: https://docs.minaprotocol.com/zkapps/how-to-write-a-zkapp-ui#enabling-coop-and-coep-headers
+/** @type {import('@sveltejs/kit').Handle} */
+export async function handle({ event, resolve }) {
+  const response = await resolve(event);
+  response.headers.set('Cross-Origin-Opener-Policy', 'same-origin');
+  response.headers.set('Cross-Origin-Embedder-Policy', 'require-corp');
+  return response;
+}
+`;
+
+export const serverHooksSvelteTypeScript = `
+import type { Handle } from "@sveltejs/kit";
+
+// To enable o1js for the web, we must set the COOP and COEP headers.
+// See here for more information: https://docs.minaprotocol.com/zkapps/how-to-write-a-zkapp-ui#enabling-coop-and-coep-headers
+export const handle: Handle = async ({ event, resolve }) => {
+  const response = await resolve(event);
+  response.headers.set('Cross-Origin-Opener-Policy', 'same-origin');
+  response.headers.set('Cross-Origin-Embedder-Policy', 'require-corp');
+  return response;
+}
+`;
